src/main/java/restaurant/security/impl/AuthenticationServiceImpl.java
package restaurant.security.impl;

import java.util.Optional;
import java.util.Set;
import org.springframework.security.crypto.password.PasswordEncoder;
import org.springframework.stereotype.Service;
import restaurant.exception.AuthenticationException;
import restaurant.model.Role;
import restaurant.model.User;
import restaurant.security.AuthenticationService;
import restaurant.service.RoleService;
import restaurant.service.ShoppingCartService;
import restaurant.service.UserService;

@Service
public class AuthenticationServiceImpl implements AuthenticationService {
    private final UserService userService;
    private final RoleService roleService;
    private final ShoppingCartService shoppingCartService;
    private final PasswordEncoder passwordEncoder;

    public AuthenticationServiceImpl(UserService userService,
                                     RoleService roleService,
                                     ShoppingCartService shoppingCartService,
                                     PasswordEncoder passwordEncoder) {
        this.userService = userService;
        this.roleService = roleService;
        this.shoppingCartService = shoppingCartService;
        this.passwordEncoder = passwordEncoder;
    }

    @Override
    public User register(String email, String password) {
        User user = userService.add(new User(email, password,
                Set.of(roleService.getByName(Role.RoleName.USER))));
        shoppingCartService.registerNewShoppingCart(user);
        return user;
    }

    @Override
    public User login(String email, String password) throws AuthenticationException {
        Optional<User> user = userService.findByEmail(email);
        if (user.isEmpty() || !passwordEncoder.matches(password, user.get().getPassword())) {
            throw new AuthenticationException("Incorrect login data.");
        }
        return user.get();
    }
}
